Split up sine utilities into ui utilities to keep the import heirarchy stable
